refactor: extract editor dialogs
